requirements.txt
aiohttp==3.8.4
fastapi==0.95.1
langchain==0.0.158
PyJWT==2.6.0
pymongo==4.3.3
python-dotenv==1.0.0
Requests==2.30.0
frozenlist==1.3.1
unstructured
openai
pdf2image
